Listening for single value events has its own class
